pyproject.toml
[tool.poetry]
name = "pretiac"
version = "0.2.0"
description = "pretiac: A PREtty Typed Icinga2 Api Client."
authors = ["Josef Friedrich <josef@friedrich.rocks>"]
readme = "README.rst"
repository = "https://github.com/Josef-Friedrich/PREtty-Typed-Icinga2-Api-Client_py"
license = "GPL-3.0-only"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2"
types-requests = "^2"
pyyaml = "^6"
pydantic = "^2"
types-pyyaml = "^6"


[tool.poetry.group.dev.dependencies]
pytest = "^8"
tox = "^4"
ruff = "^0"
sphinx = "^7"
sphinx-rtd-theme = "^2"
sphinx-tags = "^0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
docstring-code-format = true

[tool.poetry.scripts]
"icinga-api.py" = "pretiac.cli:main"
